Make static function
