+++ android/src/main/java/com/swmansion/gesturehandler/core/PanGestureHandler.kt
@@ -215,7 +215,7 @@ class PanGestureHandler(context: Context?) : GestureHandler<PanGestureHandler>()
     }
 
     if (event.getToolType(0) == MotionEvent.TOOL_TYPE_STYLUS) {
-      stylusData = GestureUtils.updateStylusData(event)
+      stylusData = StylusData.fromEvent(event)
     }
 
     val state = state
